+++ crates/rust-client/src/store/web_store/js/transactions.js
@@ -44,10 +44,14 @@ export async function getTransactions(
                 }
             }
 
+            let inputNotesArrayBuffer = await transactionRecord.inputNotes.arrayBuffer();
+            let inputNotesArray = new Uint8Array(inputNotesArrayBuffer);
+            let inputNotesBase64 = uint8ArrayToBase64(inputNotesArray);
+            transactionRecord.inputNotes = inputNotesBase64;
+
             let outputNotesArrayBuffer = await transactionRecord.outputNotes.arrayBuffer();
             let outputNotesArray = new Uint8Array(outputNotesArrayBuffer);
             let outputNotesBase64 = uint8ArrayToBase64(outputNotesArray);
-
             transactionRecord.outputNotes = outputNotesBase64;
 
             let data = {
@@ -57,7 +61,7 @@ export async function getTransactions(
                 final_account_state: transactionRecord.finalAccountState,
                 input_notes: transactionRecord.inputNotes,
                 output_notes: transactionRecord.outputNotes,
-                script_hash: transactionRecord.scriptHash ? transactionRecord.scriptHash : null,
+                tx_script_hash: transactionRecord.scriptHash ? transactionRecord.scriptHash : null,
                 tx_script: txScriptBase64,
                 block_num: transactionRecord.blockNum,
                 commit_height: transactionRecord.commitHeight ? transactionRecord.commitHeight : null
@@ -89,16 +93,13 @@ export async function insertTransactionScript(
             throw new Error("Script hash must be provided");
         }
 
-        let scriptHashArray = new Uint8Array(scriptHash);
-        let scriptHashBase64 = uint8ArrayToBase64(scriptHashArray);
-
         let txScriptBlob = null;
         if (txScript) {
             txScriptBlob = new Blob([new Uint8Array(txScript)]);
         }
 
         const data = {
-            scriptHash: scriptHashBase64,
+            scriptHash,
             txScript: txScriptBlob
         }
 
@@ -125,21 +126,17 @@ export async function insertProvenTransactionData(
     committed
 ) {
     try {
-        let scriptHashBase64 = null;
+        let inputNotesBlob = new Blob([new Uint8Array(inputNotes)]);
         let outputNotesBlob = new Blob([new Uint8Array(outputNotes)]);
-        if (scriptHash !== null) {
-            let scriptHashArray = new Uint8Array(scriptHash);
-            scriptHashBase64 = uint8ArrayToBase64(scriptHashArray);
-        }
 
         const data = {
             id: transactionId,
             accountId: accountId,
             initAccountState: initAccountState,
             finalAccountState: finalAccountState,
-            inputNotes: inputNotes,
+            inputNotes: inputNotesBlob,
             outputNotes: outputNotesBlob,
-            scriptHash: scriptHashBase64,
+            scriptHash: scriptHash ? scriptHash : null,
             blockNum: blockNum,
             commitHeight: committed ? committed : null
         }