+++ crates/rust-client/src/store/web_store/sync/js_bindings.rs
@@ -20,23 +20,23 @@ extern "C" {
     // ================================================================================================
 
     #[wasm_bindgen(js_name = addNoteTag)]
-    pub fn idxdb_add_note_tag(tags: String) -> js_sys::Promise;
+    pub fn idxdb_add_note_tag(tags: Vec<u8>) -> js_sys::Promise;
 
     #[wasm_bindgen(js_name = applyStateSync)]
     pub fn idxdb_apply_state_sync(
         block_num: String,
         nullifiers: Vec<String>,
         nullifier_block_nums: Vec<String>,
-        block_header: String,
-        chain_mmr_peaks: String,
+        block_header: Vec<u8>,
+        chain_mmr_peaks: Vec<u8>,
         has_client_notes: bool,
         serialized_node_ids: Vec<String>,
         serialized_nodes: Vec<String>,
         output_note_ids: Vec<String>,
-        output_note_inclusion_proofs: Vec<String>,
+        output_note_inclusion_proofs: Vec<u8>,
         input_note_ids: Vec<String>,
-        input_note_inclusion_proofs: Vec<String>,
-        input_note_metadatas: Vec<String>,
+        input_note_inclusion_proofs: Vec<u8>,
+        input_note_metadatas: Vec<u8>,
         transactions_to_commit: Vec<String>,
         transactions_to_commit_block_nums: Vec<String>,
     ) -> js_sys::Promise;