+++ crates/rust-client/src/store/web_store/sync/models.rs
@@ -1,6 +1,7 @@
-use alloc::string::String;
+use alloc::{string::String, vec::Vec};
 
-use serde::{Deserialize, Serialize};
+use base64::{engine::general_purpose, Engine as _};
+use serde::{de::Error, Deserialize, Deserializer, Serialize};
 
 #[derive(Serialize, Deserialize)]
 pub struct SyncHeightIdxdbObject {
@@ -9,5 +10,16 @@ pub struct SyncHeightIdxdbObject {
 
 #[derive(Serialize, Deserialize)]
 pub struct NoteTagsIdxdbObject {
-    pub tags: String,
+    #[serde(deserialize_with = "base64_to_vec_u8_required", default)]
+    pub tags: Vec<u8>,
+}
+
+fn base64_to_vec_u8_required<'de, D>(deserializer: D) -> Result<Vec<u8>, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    let base64_str: String = Deserialize::deserialize(deserializer)?;
+    general_purpose::STANDARD
+        .decode(&base64_str)
+        .map_err(|e| Error::custom(format!("Base64 decode error: {}", e)))
 }